feat: initial Python SHACL unit testing framework
Set up a fully qualified and functional pySHACL-based Pytest project,
which takes representative sample data and validates them against a set
of SHACL shapes. In this case, the data are treated as units to be
tested, representing valid and invalid states pertaining to the shapes
in test. Both the conformance and result count are checked.

To introduce new tests, create invalid and (optionally) valid data files
in `tests/test_data/shacl` with a folder and files named according to
the shape in test. Write the tests in `tests/unit/shacl` following the
same naming convention.

In a test module, adapt the `TEST_FILE_NAME` and the
`@pytest.mark.parametrize` values according to your data and
expectations (typically only the `expected_result_count` for the invalid
case needs to be updated).

One can trigger a display of the full report by triggering a test
failure, in turn by changing one of the expected values for the
parameters (e.g. simply change the `expected_result_count`). Inspect the
SHACL violation report to determine the expected result count.

Targeted tests can be programmed by accessing the complete report
object. However, this would yield an outcome similar to having
TopBraid-like DASH test cases, where one would end up with a complete
report graph that needs to be asserted against a known result graph.
Such a feature can be achieved with SPARQL ASK assertions.

A helper script `validation_runner.py` is provided to facilitate manual
one-off tests.