feat: add external filtering to extraction script for MoSCow
Add a sophisticated filtering mechanism to the entity extraction script,
with the intention to use the UML/XMI export as a lookup table to filter
on the MoSCow "qualifier", for producing coverage reports against each
qualifier. The first exercise is for MUST (mandatory), which has been
verified manually against the XMI (there are 10 occurences).

We now support the following arguments which must be provided together:

```sh
--filter-csv <FILE> --filter-column <COL> --filter-value <VAL>
```

These are optional as a group, and if provided, will look up `<FILE>`
(in CSV) and join a given class or property with a value in a specific
column, and match `<VAL>` in the `<COL>` column. An additional optional
parameter `--parent-column` defaults to `parent` for matching on the
parent class or node shape of a property. The join column itself is
currently hardcoded to `entity`.

This commit also adds a hardcoded prefix replacement logic, as there are
some cases where there is a discrepancy. Making this dynamic is not
easy, as the namespaces for the XMI are defined in a separate file. We
can add more replacements on a case-by-case basis until we find or
decide on a more sophisticated extension for handling prefixes.