feat: initial, experimental SHACL test case generator
This is an experimental attempt to generate, deterministally, valid and
invalid data test cases from SHACL shapes.

It analyzes SHACL shapes and generates corresponding data examples
structured by the shape name and validity, that can be used with any
testing framework that can take the invalid and valid pairs as
"fixtures" or examples. Output is currently a list of Gherkin scenario
outline examples.

EACH TEST CASE MAY NOT VALIDATE AS EXPECTED SO MANUAL CURATION IS NEEDED