feat: update tests for unexpected violations in current model
There is a violation for dcat:accessURL `Value is not Literal with
datatype rdfs:Resource`. This is not in line with usual SHACL
expectations as this rule can never be satisfied. Until a proper
resolution is in place, we bump up the number of expected valid and
invalid violation counts.