feat: add new coverage report target for selected data folders
Add a new Make target for producing reports by the folder, for a
selected set of examples (currently hardcoded in the variable
`REFERENCE_DATA_FOLDERS`, which can be overridden when running `make`).
This is to facilitate reporting for only a subset of the test data that
are considered to be official data samples, rather than test
units/fixtures alone. Persist these reports in a structure similar to
the overall coverage reports.

P.S: The Make target for `coverage_report` is now refactored into
`coverage-report` (with a dash). It was initially meant to indicate a
different variant vs. `test-report` which was not a variant but an
extension (reports of tests). However, that convention could lead to
confusion.