fix: faulty coverage reports due to RDF parsing errors
What we do for testing SHACL rules with invalid data fragments is not
going to sit well with scripts processing RDF and expecting valid RDF,
e.g. if there is a value type assertion, the value must conform.
Therefore `rdflib` throws errors if "-1" is an
**xsd:nonNegativeInteger**.

We eliminate these test cases for now and in the future we can introduce
`riot --validate` to capture such RDF native validation issues. Update
also all the coverage reports, which now show a good overall coverage.