refactor,ci: reorganize GitHub Action testing steps
Both the coverage overall and by-data depend on the UML XMI extraction.
At most we can make a separate target for it, but both the coverage
targets should still use it instead of relying on the user to run the
dependency first. Of course this means running both sequentially would
have a redudant run of the extraction but that's a small price to pay.